backend/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
langchain-groq
python-dotenv
pydantic
supabase
instaloader
apscheduler
